[grid-ui] Decomposing NavBar to smaller components
